sales_etl_pipeline/requirements.txt
pandas
python-dotenv
sqlalchemy
mysql-connector-python
